=Change Migration

+++ database/migrations/2023_10_19_104944_create_barang_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('barang', function (Blueprint $table) {
             $table->increments('id_barang');
-            $table->integer('id_kategori');
+            $table->bigInteger('id_kategori');
             $table->integer('id_stok');
             $table->string('nama_barang')->unique();
             $table->string('merk')->nullable();
@@ -21,6 +21,7 @@ return new class extends Migration
             $table->integer('harga_jual');
             $table->integer('stok');
             $table->timestamps();
+            
             $table->foreign('id_kategori')
                   ->references('kategori')
                   ->on('id_kategori')
